requirements.txt
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
enum34==1.1.10
fire==0.6.0
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
lxml==5.2.2
numpy==2.0.0
opencv-python-headless==4.10.0.84
pdf2docx==0.5.8
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypass==0.2.1
python-docx==1.1.2
python-sane==2.9.1
python-telegram-bot==20.8
pyudev==0.24.1
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
systemd-python==235
termcolor==2.4.0
typing_extensions==4.12.2
